change to double

+++ src/main/java/com/example/PokemonManagementSystem/model/PlayerStatus.java
@@ -14,5 +14,5 @@ public class PlayerStatus {
     private Long id;
 
     private int clicks;
-    private int gold;
+    private double gold;
 }
